// src/components/editor/types.ts
export type CanvasComponentData = ComponentType & {
  id: number;
  props: ComponentProps;
};

export type Step = {
  id: number;
  name: string;
  components: CanvasComponentData[];
  width?: string;
  spacing?: string;
  borderRadius?: string;
};

export type Funnel = {
  id: string;
  name: string;
  type: string;
  userId: string;
  steps: Step[];
};

export type EditorView = 'construtor' | 'fluxo' | 'design' | 'leads' | 'configuracoes';

export type CanvasBlock = {
    id: number;
    type: string;
    position: { x: number; y: number };
    parentId?: number | null;
    children?: CanvasBlock[];
};